+++ src/test/java/org/opentripplanner/graph_builder/module/linking/LinkingTest.java
@@ -54,8 +54,9 @@ public class LinkingTest {
             SplitterVertex sv0 = new SplitterVertex(null, "split", x + delta * splitVal, y + delta * splitVal, s0);
             SplitterVertex sv1 = new SplitterVertex(null, "split", x + delta * splitVal, y + delta * splitVal, s1);
 
-            P2<StreetEdge> sp0 = s0.splitDestructively(sv0);
-            P2<StreetEdge> sp1 = s1.splitDestructively(sv1);
+            var graph = new Graph();
+            P2<StreetEdge> sp0 = s0.splitDestructively(sv0, graph);
+            P2<StreetEdge> sp1 = s1.splitDestructively(sv1, graph);
 
             // distances expressed internally in mm so this epsilon is plenty good enough to ensure that they
             // have the same values